refactor(db): extract EntityMetadataRepository from StatsRepository
Splits the 598-line StatsRepository into two focused repositories:
time-series operations stay; entity metadata (icons, display names,
service-key mappings) moves to EntityMetadataRepository.

Closes #105
